tools/validate: skip empty lines (#167)
Some excel files may have stray characters on a very high row number,
which will cause very slow validation and a bunch of extra 'empty-value'
errors.

One potential issue with this is that the row-numbers after empty rows will be wrong. This can be addressed in #143/#148, which already will fix row numbers.