use enum for rule ids
- add RuleId enum to rules module
- change rule id values to enum values where possible

The decision to put the enum in the rules module was made to make it
easier to make new (and keep track of old) rules, since it's all in the
same place. This does cause a minor issue where I have to use string
values instead of the enum in a few extra places to avoid circular
imports (since the rules module already depend on other modules), but
that should be worth it.