log and continue, on missing cerberus rule handler
some times cerberus rules/errors are triggered that we don't have a
handler for yet. One example of this is the 'nullable' error, which can
occur depending on how the dataset is imported, where ends up containing
None/null values.

These errors were fatal, but this fix makes it a logged error instead,
so that validation can continue even when the data triggers edge cases.