get odm version from dictionary or table files
depending on the context.

This must be done because the dictionaries are the main source of the
available ODM versions, and so they should primarily be used, however,
they are not included in the package, so we need to fall back on getting
the versions from the assets generated from the dictionaries (which also
has the versions in their filenames).